LayoutTests/crypto/subtle/rsassa-pkcs1-v1_5-import-jwk-expected.txt

Test importing an RSA key for RSASSA-PKCS1-v1_5.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing a public key...
PASS publicKey.toString() is '[object Key]'
PASS publicKey.type is 'public'
PASS publicKey.usages is ['sign','verify']
PASS publicKey.algorithm.name is 'rsassa-pkcs1-v1_5'
PASS publicKey.algorithm.modulusLength is 2048
PASS byteArrayToHexString(publicKey.algorithm.publicExponent) is '[01 00 01]'
PASS publicKey.algorithm.hash.name is 'sha-256'

Importing a private key...
PASS privateKey.toString() is '[object Key]'
PASS privateKey.type is 'private'
PASS privateKey.usages is ['sign','verify']
PASS privateKey.algorithm.name is 'rsassa-pkcs1-v1_5'
PASS privateKey.algorithm.modulusLength is 2048
PASS byteArrayToHexString(privateKey.algorithm.publicExponent) is '[01 00 01]'
PASS privateKey.algorithm.hash.name is 'sha-256'
PASS successfullyParsed is true

TEST COMPLETE
